# docs/web-api-openapi.yaml
openapi: 3.0.3
info:
  title: YatraX Web API
  description: |
    RESTful API for YatraX web application.
    
    ## Authentication
    The API uses JWT token-based authentication.
    - Tokens expire in 7 days (default)
    - Include token in Authorization header: `Bearer <token>`
    
    ## Public Endpoints
    Some endpoints (like Trail Info) are publicly accessible and do not require authentication.
  version: 1.0.0
  contact:
    name: YatraX API Support
    email: support@yatrax.com
servers:
  - url: http://localhost:8000/api/web
    description: Development server
  - url: https://api.yatrax.com/api/web
    description: Production server
...
paths:
...
  /trails:
    post:
      tags:
        - Trails
      summary: Create a new trail
      description: Create a trail using GeoJSON Feature format
      operationId: createTrail
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoJSONFeature'
      responses:
        '201':
          description: Trail created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Trails
      summary: Get all trails
      description: Get all trails with pagination and filtering
      operationId: getAllTrails
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 100
        - name: fid
          in: query
          schema:
            type: integer
        - name: name
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Trails retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  trails:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeoJSONFeature'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trails/{id}:
    get:
      tags:
        - Trails
      summary: Get trail by ID
      description: Get a specific trail by its ID
      operationId: getTrailById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Trail retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoJSONFeature'
        '404':
          description: Trail not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Trails
      summary: Update trail
      description: Update an existing trail
      operationId: updateTrail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoJSONFeature'
      responses:
        '200':
          description: Trail updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trail not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Trails
      summary: Delete trail
      description: Delete a trail by ID
      operationId: deleteTrail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Trail deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          description: Trail not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trails/feature-collection:
    post:
      tags:
        - Trails
      summary: Import FeatureCollection
      description: Import multiple trails from a GeoJSON FeatureCollection
      operationId: createTrailsFromFeatureCollection
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - features
              properties:
                type:
                  type: string
                  enum: [FeatureCollection]
                features:
                  type: array
                  items:
                    $ref: '#/components/schemas/GeoJSONFeature'
      responses:
        '201':
          description: Trails imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  count:
                    type: integer
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trails/near/point:
    get:
      tags:
        - Trails
      summary: Find trails near a point
      description: Find trails near a specific point using geospatial queries
      operationId: findTrailsNear
      security: []
      parameters:
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            minimum: -180
            maximum: 180
          example: 86.9250
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            minimum: -90
            maximum: 90
          example: 27.9881
        - name: maxDistance
          in: query
          schema:
            type: number
            minimum: 0
            default: 10000
          description: Maximum distance in meters
        - name: minDistance
          in: query
          schema:
            type: number
            minimum: 0
            default: 0
          description: Minimum distance in meters
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        '200':
          description: Trails found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  trails:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeoJSONFeature'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trails/within/radius:
    get:
      tags:
        - Trails
      summary: Find trails within radius
      description: Find trails within a circular area (radius in kilometers)
      operationId: findTrailsWithinRadius
      security: []
      parameters:
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: radius
          in: query
          schema:
            type: number
            minimum: 0.1
            default: 10
          description: Radius in kilometers
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        '200':
          description: Trails found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  trails:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeoJSONFeature'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trails/within/polygon:
    post:
      tags:
        - Trails
      summary: Find trails within polygon
      description: Find trails within a specified polygon area
      operationId: findTrailsWithin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - polygon
              properties:
                polygon:
                  $ref: '#/components/schemas/GeoJSONPolygon'
      responses:
        '200':
          description: Trails found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  trails:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeoJSONFeature'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trails/intersecting:
    post:
      tags:
        - Trails
      summary: Find trails intersecting geometry
      description: Find trails that intersect with a specified geometry
      operationId: findTrailsIntersecting
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - geometry
              properties:
                geometry:
                  $ref: '#/components/schemas/GeoJSONGeometry'
      responses:
        '200':
          description: Trails found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  trails:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeoJSONFeature'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login/register endpoints
  schemas:
...
    GeoJSONFeature:
      type: object
      required:
        - type
        - geometry
      properties:
        type:
          type: string
          enum: [Feature]
          default: Feature
        geometry:
          $ref: '#/components/schemas/GeoJSONGeometry'
        properties:
          type: object
          additionalProperties: true
          example:
            name: "Everest Base Camp Trek"
            difficulty: 4
            fid: 123
    GeoJSONGeometry:
      type: object
      required:
        - type
        - coordinates
      properties:
        type:
          type: string
          enum: [Point, LineString, Polygon, MultiPoint, MultiLineString, MultiPolygon]
        coordinates:
          oneOf:
            - type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
            - type: array
              items:
                type: array
                items:
                  type: number
            - type: array
              items:
                type: array
                items:
                  type: array
                  items:
                    type: number
    GeoJSONPolygon:
      type: object
      required:
        - type
        - coordinates
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
    TrailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Trail created successfully"
        trail:
          $ref: '#/components/schemas/GeoJSONFeature'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          nullable: true